requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
pandas==1.3.3
openpyxl==3.0.7
Werkzeug==2.0.1 